[Phase 6 Step 1] Consolidate runtime VehicleBuilder into typestate VehicleBuilder
Phase 5 introduced a typestate `VehicleBuilder` in `jeod_sim` that built
a small `TypedVehicleConfig`. Step 0 of Phase 6 relocated the per-vehicle
config types (full `VehicleConfig`, `SrpModel`, `ShadowBody`,
`DerivedStateConfig`, `EarthLightingConfig`, `GeodeticConfig`,
`FrameSwitchConfig`, `SwitchSense`, `MassTreeAttachment`) and the
declarative `SimulationBuilder` to `jeod_sim`. Step 1 finishes the
consolidation: the typestate `VehicleBuilder` now produces the full
[`VehicleConfig`] and exposes every `Ready`-state method that the
runtime fluent builder offered (`.shadow()`, `.cannonball_srp()`,
`.lvlh()`, `.solar_beta()`, `.earth_lighting()`, `.external_force()`,
`.external_torque()`, `.frame_switches()`, `.integ_source()`,
`.gravity_gradient()`, `.structural_transform()`, `.orbital_elements()`,
`.euler_angles()`, `.geodetic()`).

Net effect:

- One vehicle builder, four phantom states gating required steps.
- `build()` returns the full `VehicleConfig` (drop-in for
  `SimulationBuilder::add_body`).
- Mission code: `VehicleBuilder::new().with_state(trans)
  .three_dof_point_mass(420_000.0.kg()).rk4().lvlh().build()`.
- Compile-time safety: missing `.with_state` / mass / integrator stays
  a compile error; runtime panics from half-configured builders are
  gone.

Removals:

- `TypedVehicleConfig` (Phase 5 stepping-stone) — superseded by the
  full `VehicleConfig` returned from `build()`.
- `jeod_runner::builder::VehicleBuilder` (the runtime fluent struct
  + ~200 lines of impl) — every method moved into the typestate
  `Ready` impl. `jeod_runner` re-exports the typestate version from
  `jeod_sim` so external `use jeod_runner::VehicleBuilder` continues
  to compile.
- `jeod_runner::builder` reduced to just the runner-side terminal
  methods (`Simulation::from_builder`, `SimulationBuilderExt`).

Two convenience entry points on the typestate builder:

- `VehicleBuilder::new()` → `NeedsState` (unchanged from Phase 5).
- `with_state(s: TranslationalState)` accepts the untyped struct so
  60+ existing callers using `TranslationalState` directly compile
  without churn. `with_translational(s: TranslationalStateTyped)`
  remains the typed entry point for new code.

Verification:

- `cargo fmt --check` / `cargo clippy --workspace --tests --examples
  -- -D warnings` clean.
- `cargo nextest run --workspace -E 'not test(tier3_)'` — 723 pass
  (includes the new `ready_state_full_surface` test on the typestate
  builder).
- `cargo nextest run -p jeod_runner --test tier3_sim_dyncomp_run2
  --test tier3_sim_dyncomp_run3` — 4/4 pass.
- `tier3_baseline_diff` — 76 matched, 0 violations (kernels
  bit-identical).
